Add files via upload
I fixed the names of some methods I made prediction much cleaner at the cost of some duplicate computing. Changed the Color as per Skyler's Request. The Prediction and Prediction point are now separate and look much cleaner. I removed the candlestick graph because it became to burdensome on the user to sort through the info. EVERYTHING IS TOGGLE ON/OFF ABLE!!!!! Im sorry im tired and excited. NOTE:

The Code push back the user's time selected THIS IS BECAUSE THE ALGO GETS VERRRRRRRRY INNCAURATE IF THE YEAR SPAN ISNT ENOUGH DAYS LONG